+++ config/bootstrap.php
@@ -25,6 +25,13 @@ if (!defined('DATABASE_BACKUP_EXECUTABLES')) {
     ]);
 }
 
+/**
+ * Valid extensions. Names as keys and compressions as values
+ */
+if (!defined('DATABASE_BACKUP_EXTENSIONS')) {
+    define('DATABASE_BACKUP_EXTENSIONS', ['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false]);
+}
+
 //Database connection
 if (!Configure::check('DatabaseBackup.connection')) {
     Configure::write('DatabaseBackup.connection', 'default');